move ArchiveManager to new Record namespace
